+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Changes in version 2.??.? - 2020-??-??
+
+
 # Changes in version 2.12.2 - 2020-04-30
 
  * Minor changes